refactor: 정보 입력 컴포넌트 이름 및 패키지 리팩토링
